SOLID refactoring in haddan_classes
